Fixed BroCode lesson numbers (skip lesson 1)
I just realized that "lesson 1" was just an primer on C and no actual
code. In case I come back to this in future and can't figure out why the
lesson numbers in the video don't match the lesson numbers in my git, I
have renumbered the lesson folders so that they do.